Silenced warning with cast on OS X (following identical pull request from readerwriterqueue#56)

+++ blockingconcurrentqueue.h
@@ -142,7 +142,7 @@ namespace details
 			bool timed_wait(std::uint64_t timeout_usecs)
 			{
 				mach_timespec_t ts;
-				ts.tv_sec = timeout_usecs / 1000000;
+				ts.tv_sec = static_cast<unsigned int>(timeout_usecs / 1000000);
 				ts.tv_nsec = (timeout_usecs % 1000000) * 1000;
 
 				// added in OSX 10.10: https://developer.apple.com/library/prerelease/mac/documentation/General/Reference/APIDiffsMacOSX10_10SeedDiff/modules/Darwin.html
